search test
The error you're encountering is due to a dependency resolution conflict in npm. The grunt-todos plugin you're trying to install requires a version of Grunt that's around 0.4.0, as indicated by peer grunt@"~0.4.0". However, your project is using a more recent version of Grunt (1.6.1). This version mismatch is causing npm to throw an error.

To resolve this issue, you have a few options: